checkout_git_repo: do not add checkout_dir
Do not make repo relative to checkout_dir, because our callers may
have already done so, and that creates checkouts into
checkout/checkout/repo when checkout_dir is relative.

Also, do not rebase a checkout that is a symlink.  If someone symlinks a checkout into their developer tree, we do not want to change the version they are trying to work with.

Fixes GH-61