Ignore ExistingProvider on adding container_network_config
Copy the code from PodmanPodModel.resolve_networking into
PodmanNetworkMixin that ignores an ExistingProvider adding a
NetworkConfig based on an oci_container_network_config.

This error generally gets triggered if the NetworkConfig is none or if force is True.
The first time is when a ModelGroup or layout calls resolve_networking. The second time can be when a pod, PodmanPodModel or similar comes to ready and again calls resolve_networking.