Restructure TaskWrapper.__call__ to be async
TaskWrapper.__call__ had a complex structure so it was async only when the underlying setup_task was async.
As a consequence hash functions for tasks could not be async.
In addition, the complexity would explode as we added depend_on support.

There do not appear to be a significant number of call sites outside of setup_task tests that take advantage of this functionality.

So simplify the code of TaskWrapper.__call__ by making it async all
the time.  If we run into cases where a non-async caller needs to call
a setup task, put the required functionality into a non-setup-task
function that gets called from an async setup-task.