Restructure so SetupTaskContext encompasses should_run
We need a way to keep track of whether dependencies have already been run between should_run and the actual __call__.
We do not want to modify the signature of __call__ and mostly do not want to modify the return of should_run.
So rearrange when SetupTaskContext is instantiated so that we can store the information there.

This commit only include rearranging SetupTaskContext
instantiation, and does not include the dependency state.

As a consequence, errors in should_run are introspected as belonging to the setup_task involved.