ansible_helper: do not run setup tasks
Now that container stamps no longer live in containers, bringing a
container to ready from ansible_chroot_helper will try to regenerate
the ssh key.  But the container is already running so this fails.

We never want ansible_chroot_helper to run setup_tasks because it does
not have enough Carthage config to do so safely.  So never bring the
container to ready.