All PodmanImages can be built
We had a check to try and determine if an image could be built.
Remove that check because it was broken in two respects. First it
relied on looking at whether an image had setup_tasks before
Customizations are processed.  Second, even if there are no tasks, we
might be changing the image command, entrypoint, or environment.