avoid base_image loop
Commit 16bad34d introduced a potential loop:

* An image specifies base_image by string

* oci_container_image is set in the injector hierarchy.

We synthesize a PodmanImage for the base_image, using the given base_image as the oci_image_tag for the base_image.  That is, we wish to have an image with the given tag to use as our base.
However previously we did not clear oci_container_image out of the environment or specify base_image when constructing the Synthesized PodmanImage.

The base_image of PodmanImage was a reference to oci_container_image
with _ready=None. This would tend to create a cycle in the common case
that oci_container_image pointed to ourselves.

This commit introduces two changes to break that cycle:

* base_image is _ready=False on PodmanImage.

* Explicitly set base_image to None when constructing the synthesized image.