OciImage: restore class injections
Commit a4d5ea60963f significantly restricted the set of InjectionKeys
that an image registered itself by default. This ended up being too
restrictive: it is common to define a class like DnsServerImageModel,
add-provider that class, and then expect to be able to set
oci_container_image to an injector_xref to that class.

Expand supplementary_injection_keys in OciImage to allow this to continue to work.