PodmanLocalSocket: container host for /run/podman/podman.sock
* filesystem_access_core: split out logic for setting up filesystem
access via sshfs.

* SshfsContext: state used by filesystem_access_core

* PodmanContainerHost: Add filesystem_access_container and
filesystem_access_volume entry points

* PodmanLocalSocket: a container host that uses podman --remote.  For
filesystem_access_container run podman exec and try to find an sftp
server within the container.  For filesystem_access_volume, use a
configured image, mount the volume, and try to access from that
container.

* podman_sftp_server_mount: Find the local sftp server and mount into
a container at a location
that Carthage can find. Not as useful as it could be if the container
host has a newer glibc than the container.

* Podman config schema: store image to use for volume access.

* PodmanContainer and PodmanVolume: Update to use
filesystem_access_container and filesystem_access_volume