Podman tests: use trixie
Use trixie because we are using that on dev machines and so we need
glibc 2.38 in the containers to mount in an sftp-server.  A better
long-term fix would be to use a ContainerCustomization to install
openssh-sftp-server in a base image to use.

Also, deploy the container host rather than just bringing it to ready
so it is started.