PodmanContainer.container_exec checks for running
You can only execute a command in a running container. We have seen
some CI failures where a container was not running and we tried to
perform a container exec. Both for a better error message and to try
and see if Carthage internal state is out of sync fail if self.running
is False at container_exec time.