Pull with Explicit URL
At least in common CI configurations, the URL may embed a CI token. So
the URL changes between runs.  For that reason, include the currently
configured URL in the git pull command when pulling/fetching pulugins.

As a result, if the checkout URL changes between runs, but the stem
does not, we will clone from the original URL, but fetch/pull from the
updated URL.