Convert to pyroute2 and improve mapping validation
The original motivation for this patch was to support alt_ifname
matching.

But I hope to have clarified the API by splitting the logic up a bit
differently:

match_links takes a dict of NetworkLinks and a dict of local
interfaces, and returns the best mapping.

process_local_network_config reads the local interfaces, calls
match_links to get the mapping, and then updates the NetworkLinks with
the local interface data.

The advantage of this approach is that the mapping logic is no longer
split between two functions, and that match_links can optimize the
match without being dependent on the order of calling match_link.  For
example, if there are ambiguous matches, it can give a much better
error.

Because netifaces doesn't provide alt_ifnames, and because it's no
longer maintained, I moved to import pyroute2 to read interface data.
I think we can easily update update_gateways and remove the dependency
on netifaces, but haven't done that in this patch.