set self.running in Vm.find
We want to have running be valid after find() so set self.running.
Note that find can still return True when is_machine_running() returns
False, either if the volume already exists, or if the VM is defined by
nut not started.  Also, we do not call is_machine_running because the
IPAddress find logic can be slow.