V4Config.resolve vs net.assign_addresses
v4_config.resolve needs to be called before merged_v4_config ever gets
called on a link.  The podman resolve_networking was trying to assign
addresses for the entire net during resolve_model time. That ended up
calling merged_v4_config on links where the v4_config had not yet been
resolved.

It is desirable to call assign_addresses at modeling time so that
things like proxy configuration work.  But instead of calling
assign_addresses on the entire net, when calling at resolve_model
time, only call for the link we know has fully finished.

It is still reasonable for the systemd network template to call
assign_addresses on the entire net.  That runs as setup_task time
which is generally after resolve_networking has completed globally.

The symptom of getting this ordering wrong is often things like
address_within_network get left as a function on the merged_v4_config
but are correct on the actual v4_config.