Allow sending stdin to commands in podman containers
We follow the convention in other run_command() functions of passing
the kwargs all the way through to sh.podman().

This lets us do things like:

    async def restore_from(self, p):
        async with self.machine_running():
	    with p.open('rb') as f:
                await self.run_command('bash', '-c', 'psql -q -U "$POSTGRES_USER" template1', _in=p.open('rb'))

without needing to copy the file to the container and then delete it.

I haven't tested this talking to podman on a remote container host.